Switch to /tmp as the download directory
Downloads always need to be moved anyway if they are important. Otherwise they
can be deleted. Storing them in /tmp makes this more convenient since manual
deletion is no longer required. Moving downloads to a more permanent location
will not change. This is only safe on single-user systems due to the
permissions on the /tmp directory.